// ch10/devjobs/libs/api-types/model/user.ts
/**
 * Generated by orval v7.12.2 🍺
 * Do not edit manually.
 * DevJobs API
 * API for the DevJobs platform
 * OpenAPI spec version: 1.0.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: string;
  name: string;
  email: string;
  password: string;
  role: UserRole;
  /** Required if role is companyUser */
  companyId?: string;
}